Split out x86_64 relocation code
