+++ CMakeLists.txt
@@ -47,6 +47,8 @@ target_sources("spud"
 	"src/zydis_utils.h"
 	"src/detour/x86_64/x86_64.cc"
 	"src/detour/x86_64/x86_64.h"
+    "src/detour/x86_64/relocators.cc"
+    "src/detour/x86_64/relocators.h"
 	"src/detour/aarch64/aarch64.cc"
 	"src/detour/aarch64/aarch64.h"
     "src/detour/detour.cc"
