+++ src/detour/x86_64/x86_64.cc
@@ -1,5 +1,6 @@
 #include "x86_64.h"
 #include "detour/detour_impl.h"
+#include "relocators.h"
 
 #include <spud/detour.h>
 #include <spud/memory/protection.h>
@@ -25,304 +26,14 @@
 
 // Here shall be dragons, at some point
 
-#pragma optimize("", off)
-
 namespace spud {
 namespace detail {
 namespace x64 {
-
-/*
-  jmp [rip+0]
-  0x00000000000
-*/
-constexpr auto kAbsoluteJumpSize = 6 + sizeof(uintptr_t);
-
-/*
-        push r15
-        mov r15, qword ptr [0x0]
-        cmp byte ptr [r15], 1
-        pop r15
-*/
-constexpr auto kReloCompareSize = 16 + sizeof(uintptr_t);
-constexpr auto kReloCompareExpandSize =
-    kReloCompareSize - 8 - sizeof(uintptr_t);
-
-/*
-push   r15
-mov    r15, qword ptr[0x0]
-addsd  xmm3,mmword ptr [r15]
-pop    r15
-*/
-constexpr auto kReloAddsdSize = 17 + sizeof(uintptr_t);
-constexpr auto kReloAddsdExpandSize = kReloAddsdSize - 9 - sizeof(uintptr_t);
-
-constexpr auto kReloMovzxSize = 16 + sizeof(uintptr_t);
-constexpr auto kReloMovzxExpandSize = kReloAddsdSize - 10 - sizeof(uintptr_t);
-
-constexpr auto kReloMovSize = 16 + sizeof(uintptr_t);
-constexpr auto kReloMovExpandSize = kReloAddsdSize - 9 - sizeof(uintptr_t);
 
 struct RelocationResult {
   std::vector<uint8_t> data;
   size_t copy_offset;
 };
-
-struct RelocationMeta {
-  uintptr_t size;
-  uintptr_t expand;
-  std::function<void(uintptr_t, const RelocationEntry &,
-                     asmjit::x86::Assembler &)>
-      gen_relo_data;
-  std::function<uintptr_t(
-      uintptr_t trampoline_address, uintptr_t trampoline_start,
-      uintptr_t target_start, size_t target_offset, const RelocationEntry &relo,
-      uintptr_t relocation_data, asmjit::x86::Assembler &assembler)>
-      gen_relo_code;
-};
-
-struct ReloInstruction {
-  ZydisMnemonic i = ZYDIS_MNEMONIC_INVALID;
-  ZydisBranchType b = ZYDIS_BRANCH_TYPE_NONE;
-
-  constexpr ReloInstruction(ZydisMnemonic mn, ZydisBranchType b)
-      : i(mn), b(b) {}
-  constexpr ReloInstruction(ZydisMnemonic mn) : i(mn) {}
-  constexpr ReloInstruction(ZydisDecodedInstruction instruction) {
-    if (instruction.meta.branch_type != ZYDIS_BRANCH_TYPE_NONE ||
-        instruction.mnemonic == ZYDIS_MNEMONIC_JMP) {
-      i = ZYDIS_MNEMONIC_INVALID;
-      b = ZYDIS_BRANCH_TYPE_MAX_VALUE;
-    } else {
-      i = instruction.mnemonic;
-    }
-  }
-
-  bool operator==(const ReloInstruction &other) const {
-    return (this->i == other.i && this->b == other.b);
-  }
-};
-
-constexpr ReloInstruction JUMP_RELO_JMP_INSTRUCTION = {
-    ZYDIS_MNEMONIC_INVALID, ZYDIS_BRANCH_TYPE_MAX_VALUE};
-
-struct ReloInstructionHasher {
-  std::size_t operator()(const ReloInstruction &relo) const {
-    return relo.i << 16 | relo.b;
-  }
-};
-
-const std::unordered_map<ReloInstruction, RelocationMeta, ReloInstructionHasher>
-    relo_meta = {
-        {JUMP_RELO_JMP_INSTRUCTION,
-         {.size = sizeof(uintptr_t),
-          .expand = 0,
-          .gen_relo_data =
-              [](auto target_start, auto &relo, auto &assembler) {
-                using namespace asmjit;
-                using namespace asmjit::x86;
-
-                const auto jump_target = (ZyanI64)target_start +
-                                         (ZyanI64)relo.address +
-                                         (relo.instruction.raw.imm[0].value.s +
-                                          (ZyanI64)relo.instruction.length);
-
-                assembler.jmp(ptr(rip, 0));
-                assembler.embed(&jump_target, sizeof(jump_target));
-              },
-          .gen_relo_code =
-              [](uintptr_t trampoline_address, uintptr_t trampoline_start,
-                 uintptr_t target_start, size_t target_offset,
-                 const RelocationEntry &relo, uintptr_t relocation_data,
-                 asmjit::x86::Assembler &assembler) {
-                auto *code = assembler.code();
-                auto *code_data = code->textSection()->data();
-
-                auto target_code = reinterpret_cast<uint8_t *>(
-                    target_offset + code_data + trampoline_start +
-                    relo.address + relo.instruction.raw.imm[0].offset);
-
-                auto jump_target = relocation_data + trampoline_start -
-                                   relo.instruction.length - relo.address -
-                                   target_offset;
-
-                switch (relo.instruction.raw.imm[0].size) {
-                case 8: {
-                  assert(jump_target < 0xFF &&
-                         "immediate size too small for relocation");
-                  *(int8_t *)(target_code) = jump_target;
-                } break;
-                case 16: {
-                  *(int16_t *)(target_code) = jump_target;
-                } break;
-                case 32: {
-                  *(int32_t *)(target_code) = jump_target;
-                } break;
-                case 64: {
-                  *(int64_t *)(target_code) = jump_target;
-                } break;
-                }
-                return target_offset;
-              }}},
-        {ZYDIS_MNEMONIC_CMP,
-         {.size = kReloCompareSize,
-          .expand = kReloCompareExpandSize,
-          .gen_relo_data =
-              [](auto target_start, auto &relo, auto &assembler) {
-                const auto lea_target =
-                    target_start + relo.address +
-                    (relo.instruction.raw.disp.value + relo.instruction.length);
-                assembler.embed(&lea_target, sizeof(lea_target));
-              },
-          .gen_relo_code =
-              [](uintptr_t trampoline_address, uintptr_t trampoline_start,
-                 uintptr_t target_start, size_t target_offset,
-                 const RelocationEntry &relo, uintptr_t relocation_data,
-                 asmjit::x86::Assembler &assembler) {
-                using namespace asmjit;
-                using namespace asmjit::x86;
-
-                auto relo_lea_target = trampoline_start + relocation_data;
-                assembler.push(r15);
-                assembler.mov(r15, qword_ptr(relo_lea_target));
-                assembler.cmp(byte_ptr(r15),
-                              relo.operands[1].imm.is_signed
-                                  ? relo.operands[1].imm.value.s
-                                  : relo.operands[1].imm.value.u);
-                assembler.pop(r15);
-                target_offset += kReloCompareExpandSize;
-                return target_offset;
-              }}},
-        {ZYDIS_MNEMONIC_LEA,
-         {.size = sizeof(uintptr_t),
-          .expand = 0,
-          .gen_relo_data =
-              [](auto target_start, auto &relo, auto &assembler) {
-                const auto lea_target =
-                    target_start + relo.address +
-                    (relo.instruction.raw.disp.value + relo.instruction.length);
-                assembler.embed(&lea_target, sizeof(lea_target));
-              },
-          .gen_relo_code =
-              [](uintptr_t trampoline_address, uintptr_t trampoline_start,
-                 uintptr_t target_start, size_t target_offset,
-                 const RelocationEntry &relo, uintptr_t relocation_data,
-                 asmjit::x86::Assembler &assembler) {
-                auto *code = assembler.code();
-                auto *code_data = code->textSection()->data();
-
-                auto target_code = reinterpret_cast<uint8_t *>(
-                    target_offset + code_data + trampoline_start +
-                    relo.address);
-                auto lea_target = reinterpret_cast<uint8_t *>(
-                    target_code + relo.instruction.raw.disp.offset);
-
-                // Turn the lea into a mov of a qword ptr, we'll load the
-                // address from our data section below the instructions
-                *(target_code + 0x1) = 0x8B;
-
-                auto relo_lea_target = relocation_data + trampoline_start -
-                                       relo.instruction.length - relo.address -
-                                       target_offset;
-
-                switch (relo.instruction.raw.disp.size) {
-                case 8: {
-                  assert(relo_lea_target < 0xFF &&
-                         "displacement size too small for relocation");
-                  *(int8_t *)(lea_target) = relo_lea_target;
-                } break;
-                case 16: {
-                  *(int16_t *)(lea_target) = relo_lea_target;
-                } break;
-                case 32: {
-                  *(int32_t *)(lea_target) = relo_lea_target;
-                } break;
-                case 64: {
-                  *(int64_t *)(lea_target) = relo_lea_target;
-                } break;
-                }
-                return target_offset;
-              }}},
-        {ZYDIS_MNEMONIC_ADDSD,
-         {.size = kReloAddsdSize,
-          .expand = kReloAddsdExpandSize,
-          .gen_relo_data =
-              [](auto target_start, auto &relo, auto &assembler) {
-                const auto lea_target =
-                    target_start + relo.address +
-                    (relo.instruction.raw.disp.value + relo.instruction.length);
-                assembler.embed(&lea_target, sizeof(lea_target));
-              },
-          .gen_relo_code =
-              [](uintptr_t trampoline_address, uintptr_t trampoline_start,
-                 uintptr_t target_start, size_t target_offset,
-                 const RelocationEntry &relo, uintptr_t relocation_data,
-                 asmjit::x86::Assembler &assembler) {
-                using namespace asmjit;
-                using namespace asmjit::x86;
-
-                auto relo_lea_target = trampoline_address + relocation_data;
-                assembler.push(r15);
-                assembler.mov(r15, qword_ptr(relo_lea_target));
-                assembler.addsd(
-                    zydis_xmm_reg_to_asmjit(relo.operands[0].reg.value),
-                    qword_ptr(r15));
-                assembler.pop(r15);
-                target_offset += kReloAddsdExpandSize;
-                return target_offset;
-              }}},
-        {ZYDIS_MNEMONIC_MOV,
-         {.size = kReloMovSize,
-          .expand = kReloMovExpandSize,
-          .gen_relo_data =
-              [](auto target_start, auto &relo, auto &assembler) {
-                const auto lea_target = target_start + ZyanI64(relo.address) +
-                                        relo.instruction.length +
-                                        relo.operands[0].mem.disp.value;
-                assembler.embed(&lea_target, sizeof(lea_target));
-              },
-          .gen_relo_code =
-              [](uintptr_t trampoline_address, uintptr_t trampoline_start,
-                 uintptr_t target_start, size_t target_offset,
-                 const RelocationEntry &relo, uintptr_t relocation_data,
-                 asmjit::x86::Assembler &assembler) {
-                using namespace asmjit;
-                using namespace asmjit::x86;
-
-                auto relo_lea_target = trampoline_address + relocation_data;
-                assembler.push(r15);
-                assembler.mov(r15, qword_ptr(relo_lea_target));
-                assembler.mov(byte_ptr(r15), relo.operands[1].imm.value.s);
-                assembler.pop(r15);
-                target_offset += kReloAddsdExpandSize;
-                return target_offset;
-              }}},
-        {ZYDIS_MNEMONIC_MOVZX,
-         {.size = kReloMovzxSize,
-          .expand = kReloMovzxExpandSize,
-          .gen_relo_data =
-              [](auto target_start, auto &relo, auto &assembler) {
-                const auto lea_target =
-                    target_start + relo.address +
-                    (relo.instruction.raw.disp.value + relo.instruction.length);
-                assembler.embed(&lea_target, sizeof(lea_target));
-              },
-          .gen_relo_code =
-              [](uintptr_t trampoline_address, uintptr_t trampoline_start,
-                 uintptr_t target_start, size_t target_offset,
-                 const RelocationEntry &relo, uintptr_t relocation_data,
-                 asmjit::x86::Assembler &assembler) {
-                using namespace asmjit;
-                using namespace asmjit::x86;
-
-                auto relo_lea_target = trampoline_address + relocation_data;
-                assembler.push(r15);
-                assembler.mov(r15, qword_ptr(relo_lea_target));
-                assembler.mov(zydis_d_reg_to_asmjit(relo.operands[0].reg.value),
-                              byte_ptr(r15));
-                assembler.pop(r15);
-                target_offset += kReloMovzxExpandSize;
-                return target_offset;
-              }}}};
 
 static RelocationResult
 do_far_relocations(std::span<uint8_t> target, uintptr_t trampoline_address,
@@ -401,6 +112,7 @@ std::tuple<RelocationInfo, size_t> collect_relocations(uintptr_t address,
   uintptr_t relocation_offset = 0x0;
 
   std::vector<RelocationEntry> relocations;
+  relocations.reserve(10);
 
   constexpr auto kRuntimeAddress = 0x0;
 
@@ -456,9 +168,9 @@ size_t get_trampoline_size(std::span<uint8_t> target,
 
   auto target_start = reinterpret_cast<uintptr_t>(target.data());
   for (const auto &relot : relocation_info.relocations) {
-    auto &relo = std::get<RelocationEntry>(relot);
+    const auto &relo = std::get<RelocationEntry>(relot);
+    const auto &r_meta = relo_meta.at(relo.instruction);
 
-    auto &r_meta = relo_meta.at(relo.instruction);
     required_space += r_meta.size;
   }
   required_space += kAbsoluteJumpSize;
@@ -521,8 +233,9 @@ static RelocationResult do_far_relocations(
   size_t data_offset = target.size() + kAbsoluteJumpSize;
 
   for (const auto &relot : relocation_info.relocations) {
-    auto &relo = std::get<x64::RelocationEntry>(relot);
-    auto &r_meta = relo_meta.at(relo.instruction);
+    const auto &relo = std::get<x64::RelocationEntry>(relot);
+    const auto &r_meta = relo_meta.at(relo.instruction);
+
     data_offset += r_meta.expand;
   }
 
@@ -531,11 +244,11 @@ static RelocationResult do_far_relocations(
   x86::Assembler reloc_assembler(&reloc_code);
 
   for (const auto &relot : relocation_info.relocations) {
-    auto &relo = std::get<x64::RelocationEntry>(relot);
+    const auto &relo = std::get<x64::RelocationEntry>(relot);
+    const auto &r_meta = relo_meta.at(relo.instruction);
+
     relocation_data.emplace_back(data_offset +
                                  reloc_code.textSection()->bufferSize());
-
-    auto &r_meta = relo_meta.at(relo.instruction);
     r_meta.gen_relo_data(target_start, relo, reloc_assembler);
   }
 
@@ -545,9 +258,9 @@ static RelocationResult do_far_relocations(
   // right location in the buffer :)
   size_t target_offset = 0;
   for (const auto &relot : relocation_info.relocations) {
-    auto &relo = std::get<x64::RelocationEntry>(relot);
+    const auto &relo = std::get<x64::RelocationEntry>(relot);
+    const auto &r_meta = relo_meta.at(relo.instruction);
 
-    auto &r_meta = relo_meta.at(relo.instruction);
     if (r_meta.expand == 0) {
       assembler.embed(target.data() + copy_offset,
                       relo.address + relo.instruction.length - copy_offset);
